put the home game cards and fetching logic into separate component
