+++ src/components/Home/Cards.module.scss
@@ -1,0 +1,38 @@
+.cardsWrapper{
+  .loader{
+    @apply w-full flex justify-center visible transition-opacity duration-500;
+
+    &.loaded{
+      @apply opacity-0 hidden;
+    }
+  }
+
+  .cards{
+    display: flex;
+    margin-left: -30px; /* gutter size offset */
+    width: auto;
+    @apply opacity-0 pt-6 transition-opacity duration-500;
+
+    &.cardsFetched{
+      @apply opacity-100;
+    }
+
+    .cardsColumn{
+      padding-left: 30px; /* gutter size */
+      background-clip: padding-box;
+
+      & > div {
+        margin-bottom: 30px;
+      }
+    }
+  }
+
+  .hiddenLoadMore{
+    @apply my-5 h-10;
+  }
+
+  .loadingMore{
+    @apply fixed w-[100vw] top-0 left-0 h-[100vh] flex justify-center items-center bg-white/10 rounded-[6px] py-4;
+  }
+}
+
